added orientation and label placement to the rest of controls and split up meta files
